Email handler and builder logic
I'm taking a break, but I've started work on a emailHandler which takes in commonly used information and goes throught the logic for:

1. getting email important information(user's name, event information)
2. Generating the appropriate email type based on an incomming emailType identifier
3. calling the sendEmails function.

NOTE: currently, sendEmails only rakes in an Array of emails so single sending events must be inputted as an array. This is something I'm looking to change before the end of the sprint.